Move Package Managers importing script to Loader.py, move manager files
