+++ wingetui/PackageEngine/Managers/winget-cli_arm64/resources.pri
