+++ wingetui/PackageEngine/Managers/winget-cli_x64/resources.pri
