+++ wingetui/PackageEngine/Managers/winget.py
@@ -10,7 +10,7 @@ import subprocess
 
 from wingetui.Core.Tools import *
 from wingetui.Core.Tools import _
-from wingetui.PackageManagers.PackageClasses import *
+from wingetui.PackageEngine.Classes import *
 
 
 class WingetPackageManager(PackageManagerWithSources):
